refactor: Update data.js to use let instead of const for words array
Update the data.js file to change the declaration of the words array from const to let. This change allows for modifying the array later in the code.

Refactor the code structure and readability in index.html

Improve the code structure and readability in the index.html file. This includes adding a new CSS class called "blurred" to apply a blur effect to certain elements when clicked. Additionally, the audio button's click event now stops the propagation of the event to prevent unwanted behavior.